Limit cabal and base versions
The dependency 'setup-depends: 'Cabal' does not specify an upper bound on the version number. Each major release of the 'Cabal' package changes the API in various ways and most packages will need some changes to compile with it. If you are not sure what upper bound to use then use the next major version.